requirements.txt
flask==3.0.0
google-generativeai==0.8.3
python-docx==1.1.2
pandas==2.1.4
gunicorn==21.2.0
Werkzeug==3.0.1
